requirements.txt
torch==2.0.0            # 具體版本可根據你當前安裝的版本調整
torchvision==0.15.0     # 用於圖像數據處理和模型
numpy==1.21.5           # 用於數據處理
Pillow==9.2.0           # 用於處理圖像（PIL）
scikit-learn==1.0.2     # 用於計算分類報告和混淆矩陣

